+++ core/chains/evm/client/pool_test.go
@@ -16,12 +16,25 @@ import (
 	"github.com/tidwall/gjson"
 	"go.uber.org/zap"
 
+	"github.com/smartcontractkit/chainlink/core/chains/evm/client"
 	evmclient "github.com/smartcontractkit/chainlink/core/chains/evm/client"
 	evmmocks "github.com/smartcontractkit/chainlink/core/chains/evm/mocks"
 	"github.com/smartcontractkit/chainlink/core/internal/cltest"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils"
 	"github.com/smartcontractkit/chainlink/core/logger"
 )
+
+type poolConfig struct {
+	selectionMode string
+}
+
+func (c poolConfig) NodeSelectionMode() string {
+	return c.selectionMode
+}
+
+var defaultConfig client.PoolConfig = &poolConfig{
+	selectionMode: client.NodeSelectionMode_HighestHead,
+}
 
 func TestPool_Dial(t *testing.T) {
 	t.Parallel()
@@ -144,7 +157,7 @@ func TestPool_Dial(t *testing.T) {
 			for i, n := range test.sendNodes {
 				sendNodes[i] = n.newSendOnlyNode(t, test.sendNodeChainID)
 			}
-			p := evmclient.NewPool(logger.TestLogger(t), nodes, sendNodes, test.poolChainID)
+			p := evmclient.NewPool(logger.TestLogger(t), defaultConfig, nodes, sendNodes, test.poolChainID)
 			err := p.Dial(ctx)
 			if test.errStr != "" {
 				require.Error(t, err)
@@ -223,7 +236,7 @@ func TestUnit_Pool_RunLoop(t *testing.T) {
 	nodes := []evmclient.Node{n1, n2, n3}
 
 	lggr, observedLogs := logger.TestLoggerObserved(t, zap.ErrorLevel)
-	p := evmclient.NewPool(lggr, nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
+	p := evmclient.NewPool(lggr, defaultConfig, nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
 
 	n1.On("String").Maybe().Return("n1")
 	n2.On("String").Maybe().Return("n2")
@@ -299,7 +312,7 @@ func TestUnit_Pool_BatchCallContextAll(t *testing.T) {
 		sendonlys = append(sendonlys, s)
 	}
 
-	p := evmclient.NewPool(logger.TestLogger(t), nodes, sendonlys, &cltest.FixtureChainID)
+	p := evmclient.NewPool(logger.TestLogger(t), defaultConfig, nodes, sendonlys, &cltest.FixtureChainID)
 
 	p.BatchCallContextAll(ctx, b)
 }
@@ -332,7 +345,7 @@ func TestUnit_Pool_getBestNode(t *testing.T) {
 		nodes = append(nodes, node)
 	}
 
-	p := evmclient.NewPool(logger.TestLogger(t), nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
+	p := evmclient.NewPool(logger.TestLogger(t), defaultConfig, nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
 
 	nonce, err := p.NonceAt(ctx, address, blockNumber)
 	assert.NoError(t, err)
@@ -345,7 +358,7 @@ func TestUnit_Pool_getBestNode(t *testing.T) {
 		node.On("LatestReceivedBlockNumber").Return(int64(2))
 		nodes = append(nodes, node)
 
-		p := evmclient.NewPool(logger.TestLogger(t), nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
+		p := evmclient.NewPool(logger.TestLogger(t), defaultConfig, nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
 
 		nonce, err := p.NonceAt(ctx, address, blockNumber)
 		assert.NoError(t, err)
@@ -360,7 +373,7 @@ func TestUnit_Pool_getBestNode(t *testing.T) {
 		node.On("LatestReceivedBlockNumber").Return(int64(3))
 		nodes = append(nodes, node)
 
-		p := evmclient.NewPool(logger.TestLogger(t), nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
+		p := evmclient.NewPool(logger.TestLogger(t), defaultConfig, nodes, []evmclient.SendOnlyNode{}, &cltest.FixtureChainID)
 
 		nonce, err := p.NonceAt(ctx, address, blockNumber)
 		assert.NoError(t, err)