+++ core/chains/evm/config/v2_helpers_test.go
@@ -87,6 +87,7 @@ func (set chainSpecificConfigDefaultSet) asV2() v2.Chain {
 			NoNewHeadsThreshold:  models.MustNewDuration(set.nodeDeadAfterNoNewHeadersThreshold),
 			PollFailureThreshold: ptr(set.nodePollFailureThreshold),
 			PollInterval:         models.MustNewDuration(set.nodePollInterval),
+			SelectionMode:        ptr(set.nodeSelectionMode),
 		},
 		OCR: &v2.OCR{
 			ContractConfirmations:              ptr(set.ocrContractConfirmations),